pyproject.toml
[tool.poetry]
name = "patricks-resume"
version = "0.1.0"
description = ""
authors = ["Patrick Rodrigues"]

[tool.poetry.dependencies]
python = "^3.10"
google-api-python-client = "^2.42.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.1"
fastapi = "^0.75.0"
uvicorn = "^0.17.6"
python-dotenv = "^0.20.0"
PyGithub = "^1.55"
GitPython = "^3.1.27"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.1.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 80

[tool.mypy]
files = "app"
ignore_missing_imports = true
plugins = "pydantic.mypy"

[tool.black]
line-length = 80
target-version = ['py36', 'py37', 'py38', 'py39', 'py310']
include = '\.pyi?$'
exclude = [
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".tox",
    ".venv",
    "_build",
    "buck-out",
    "build",
    "dist",
    "blib2to3",
    "tests/data",
    "profiling"]
